+++ Platform/Client/httpInterface.js
@@ -43,611 +43,12 @@ exports.newHttpInterface = function newHttpInterface() {
             let requestPath = requestPathAndParameters[0].split('/')
             let endpointOrFile = requestPath[1]
 
+            // migrating switch statements to map of route name & function 
+            const routeMap = require('./Http-Routes/index').newHttpRoutes()
+            const routeFunc = routeMap.get(endpointOrFile)
+            routeFunc(httpRequest, httpResponse)
+
             switch (endpointOrFile) {
-                case 'Environment': {
-                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.env), httpResponse)
-                }
-                    break
-                case 'Bitcoin-Factory': {
-                    SA.projects.foundations.utilities.httpRequests.getRequestBody(httpRequest, httpResponse, processRequest)
-
-                    function processRequest(body) {
-                        try {
-                            if (body === undefined) {
-                                return
-                            }
-                            let params = JSON.parse(body)
-
-                            switch (params.method) {
-                                case 'updateForecastedCandles': {
-
-                                    let serverResponse = PL.servers.BITCOIN_FACTORY_SERVER.updateForecastedCandles(
-                                        params.forcastedCandles
-                                    )
-
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(serverResponse), httpResponse)
-                                    break
-                                }
-                                case 'getTestClientInstanceId': {
-
-                                    let serverResponse = PL.servers.BITCOIN_FACTORY_SERVER.getTestClientInstanceId(
-                                        params.networkCodeName,
-                                        params.userProfile,
-                                        params.clientName
-                                    )
-
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(serverResponse), httpResponse)
-                                    break
-                                }
-                                case 'getUserProfileFilesList': {
-
-                                    let serverResponse = PL.servers.BITCOIN_FACTORY_SERVER.getUserProfileFilesList(
-                                    )
-
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(serverResponse), httpResponse)
-                                    break
-                                }
-                                case 'getUserProfileFile': {
-
-                                    let serverResponse = PL.servers.BITCOIN_FACTORY_SERVER.getUserProfileFile(
-                                        params.fileName
-                                    )
-
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(serverResponse), httpResponse)
-                                    break
-                                }
-                                case 'getIndicatorFile': {
-
-                                    let serverResponse = PL.servers.BITCOIN_FACTORY_SERVER.getIndicatorFile(
-                                        params.dataMine,
-                                        params.indicator,
-                                        params.product,
-                                        params.exchange,
-                                        params.baseAsset,
-                                        params.quotedAsset,
-                                        params.dataset,
-                                        params.timeFrameLabel
-                                    )
-
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(serverResponse), httpResponse)
-                                    break
-                                }
-                                default: {
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify({ error: 'Method ' + params.method + ' is invalid.' }), httpResponse)
-                                }
-                            }
-                        } catch (err) {
-                            console.log((new Date()).toISOString(), '[ERROR] httpInterface -> Bitcoin-Factory -> Method call produced an error.')
-                            console.log((new Date()).toISOString(), '[ERROR] httpInterface -> Bitcoin-Factory -> err.stack = ' + err.stack)
-                            console.log((new Date()).toISOString(), '[ERROR] httpInterface -> Bitcoin-Factory -> Params Received = ' + body)
-
-                            let error = {
-                                result: 'Fail Because',
-                                message: err.message,
-                                stack: err.stack
-                            }
-                            SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(error), httpResponse)
-                        }
-                    }
-                }
-                    break
-                case 'WEB3': {
-                    SA.projects.foundations.utilities.httpRequests.getRequestBody(httpRequest, httpResponse, processRequest)
-
-                    async function processRequest(body) {
-                        try {
-                            if (body === undefined) {
-                                return
-                            }
-                            let params = JSON.parse(body)
-
-                            switch (params.method) {
-                                case 'getNetworkClientStatus': {
-
-                                    let serverResponse = await PL.servers.WEB3_SERVER.getNetworkClientStatus(
-                                        params.host,
-                                        params.port,
-                                        params.interface
-                                    )
-
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(serverResponse), httpResponse)
-                                    return
-                                }
-                                case 'createWalletAccount': {
-
-                                    let serverResponse = await PL.servers.WEB3_SERVER.createWalletAccount(
-                                        params.entropy
-                                    )
-
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(serverResponse), httpResponse)
-                                    return
-                                }
-                                case 'getUserWalletBalance': {
-
-                                    let serverResponse = await PL.servers.WEB3_SERVER.getUserWalletBalance(
-                                        params.chain,
-                                        params.walletAddress,
-                                        params.contractAddress
-                                    )
-
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(serverResponse), httpResponse)
-                                    return
-                                }
-                                case 'getLPTokenBalance': {
-
-                                    let serverResponse = await PL.servers.WEB3_SERVER.getLPTokenBalance(
-                                        params.chain,
-                                        params.contractAddressSA,
-                                        params.contractAddressLP
-                                    )
-
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(serverResponse), httpResponse)
-                                    return
-                                }
-                                case 'getWalletBalances': {
-
-                                    let serverResponse = await PL.servers.WEB3_SERVER.getWalletBalances(
-                                        params.host,
-                                        params.port,
-                                        params.interface,
-                                        params.walletDefinition
-                                    )
-
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(serverResponse), httpResponse)
-                                    return
-                                }
-                                case 'signData': {
-
-                                    let serverResponse = await PL.servers.WEB3_SERVER.signData(
-                                        params.privateKey,
-                                        params.data
-                                    )
-
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(serverResponse), httpResponse)
-                                    return
-                                }
-                                case 'hashData': {
-
-                                    let serverResponse = await PL.servers.WEB3_SERVER.hashData(
-                                        params.data
-                                    )
-
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(serverResponse), httpResponse)
-                                    return
-                                }
-                                case 'recoverAddress': {
-
-                                    let serverResponse = await PL.servers.WEB3_SERVER.recoverAddress(
-                                        params.signature
-                                    )
-
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(serverResponse), httpResponse)
-                                    return
-                                }
-                                case 'recoverWalletAddress': {
-
-                                    let serverResponse = await PL.servers.WEB3_SERVER.recoverWalletAddress(
-                                        params.signature,
-                                        params.account,
-                                        params.data
-                                    )
-
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(serverResponse), httpResponse)
-                                    return
-                                }
-                                case 'mnemonicToPrivateKey': {
-
-                                    let serverResponse = await PL.servers.WEB3_SERVER.mnemonicToPrivateKey(
-                                        params.mnemonic
-                                    )
-
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(serverResponse), httpResponse)
-                                    return
-                                }
-                                default: {
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify({ error: 'Method ' + params.method + ' is invalid.' }), httpResponse)
-                                }
-                            }
-                        } catch (err) {
-                            console.log((new Date()).toISOString(), '[ERROR] httpInterface -> WEB3s -> Method call produced an error.')
-                            console.log((new Date()).toISOString(), '[ERROR] httpInterface -> WEB3s -> err.stack = ' + err.stack)
-                            console.log((new Date()).toISOString(), '[ERROR] httpInterface -> WEB3s -> Params Received = ' + body)
-
-                            let error = {
-                                result: 'Fail Because',
-                                message: err.message,
-                                stack: err.stack
-                            }
-                            SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(error), httpResponse)
-                        }
-                    }
-                }
-                    break
-                case 'CCXT': {
-                    SA.projects.foundations.utilities.httpRequests.getRequestBody(httpRequest, httpResponse, processRequest)
-
-                    async function processRequest(body) {
-                        try {
-                            if (body === undefined) {
-                                return
-                            }
-                            let params = JSON.parse(body)
-
-                            switch (params.method) {
-                                case 'fetchMarkets': {
-
-                                    const exchangeClass = SA.nodeModules.ccxt[params.exchangeId]
-                                    const exchangeConstructorParams = {
-                                        'timeout': 30000,
-                                        'enableRateLimit': true,
-                                        verbose: false
-                                    }
-
-                                    let ccxtExchange = new exchangeClass(exchangeConstructorParams)
-                                    let ccxtMarkets = []
-
-                                    if (ccxtExchange.has.fetchMarkets === true) {
-                                        ccxtMarkets = await ccxtExchange.fetchMarkets()
-                                    }
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(ccxtMarkets), httpResponse)
-                                    return
-                                }
-                                case 'listExchanges': {
-                                    let exchanges = []
-                                    for (let i = 0; i < SA.nodeModules.ccxt.exchanges.length; i++) {
-                                        let exchangeId = SA.nodeModules.ccxt.exchanges[i]
-
-                                        const exchangeClass = SA.nodeModules.ccxt[exchangeId]
-                                        const exchangeConstructorParams = {
-                                            'timeout': 30000,
-                                            'enableRateLimit': true,
-                                            verbose: false
-                                        }
-                                        let ccxtExchange
-                                        try {
-                                            ccxtExchange = new exchangeClass(exchangeConstructorParams)
-                                        } catch (err) {
-                                        }
-                                        if (ccxtExchange === undefined) {
-                                            continue
-                                        }
-
-
-                                        if (ccxtExchange.has.fetchOHLCV === params.has.fetchOHLCV) {
-                                            if (ccxtExchange.has.fetchMarkets === params.has.fetchMarkets) {
-                                                if (ccxtExchange.timeframes['1m'] !== undefined) {
-                                                    let exchange = {
-                                                        name: ccxtExchange.name,
-                                                        id: ccxtExchange.id
-                                                    }
-                                                    exchanges.push(exchange)
-                                                }
-                                            }
-                                        }
-                                    }
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(exchanges), httpResponse)
-                                    return
-                                }
-                            }
-
-                            let content = {
-                                err: global.DEFAULT_FAIL_RESPONSE // method not supported
-                            }
-                            SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(content), httpResponse)
-
-                        } catch (err) {
-                            console.log((new Date()).toISOString(), '[INFO] httpInterface -> CCXT FetchMarkets -> Could not fetch markets.')
-                            let error = {
-                                result: 'Fail Because',
-                                message: err.message
-                            }
-                            SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(error), httpResponse)
-                        }
-                    }
-                }
-                    break
-                case 'DEX':
-                    switch (requestPath[2]) {
-                        case 'CreateNewWallet':
-                            console.log('creating new wallet')
-                            let dexWallet = SA.projects.decentralizedExchanges.modules.wallets.newDecentralizedExchangesModulesWallets()
-                            dexWallet.initialize()
-                                .then(() => {
-                                    dexWallet.createWallet()
-                                        .then(wallet => {
-                                            responseBody = JSON.stringify({
-                                                address: wallet.address,
-                                                mnemonic: wallet.mnemonic.phrase,
-                                                privateKey: wallet.privateKey,
-                                                publicKey: wallet.publicKey
-                                            })
-                                            SA.projects.foundations.utilities.httpResponses.respondWithContent(responseBody, httpResponse)
-                                        })
-                                        .catch(err => {
-                                            console.error(err)
-                                            SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                        })
-                                })
-                                .catch(err => {
-                                    console.error(err)
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                })
-                            break
-                        case 'ImportWalletFromMnemonic':
-                            console.log('importing wallet from mnemonic')
-                            SA.projects.foundations.utilities.httpRequests.getRequestBodyAsync(httpRequest, httpResponse)
-                                .then(body => {
-                                    let config = JSON.parse(body)
-                                    let dexWallet = SA.projects.decentralizedExchanges.modules.wallets.newDecentralizedExchangesModulesWallets()
-                                    dexWallet.initialize()
-                                        .then(() => {
-                                            dexWallet.importWalletFromMnemonic(config.mnemonic)
-                                                .then(wallet => {
-                                                    responseBody = JSON.stringify({
-                                                        address: wallet.address
-                                                    })
-                                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(responseBody, httpResponse)
-                                                })
-                                                .catch(err => {
-                                                    console.error(err)
-                                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                                })
-                                        })
-                                        .catch(err => {
-                                            console.error(err)
-                                            SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                        })
-                                })
-                                .catch(err => {
-                                    console.error(err)
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                })
-                            break
-                        case 'ImportWalletFromPrivateKey':
-                            console.log('importing wallet from private key')
-                            SA.projects.foundations.utilities.httpRequests.getRequestBodyAsync(httpRequest, httpResponse)
-                                .then(body => {
-                                    let config = JSON.parse(body)
-                                    let dexWallet = SA.projects.decentralizedExchanges.modules.wallets.newDecentralizedExchangesModulesWallets()
-                                    dexWallet.initialize()
-                                        .then(() => {
-                                            dexWallet.importWalletFromPrivateKey(config.privateKey)
-                                                .then(wallet => {
-                                                    responseBody = JSON.stringify({
-                                                        address: wallet.address,
-                                                        publicKey: wallet.publicKey
-                                                    })
-                                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(responseBody, httpResponse)
-                                                })
-                                                .catch(err => {
-                                                    console.error(err)
-                                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                                })
-                                        })
-                                        .catch(err => {
-                                            console.error(err)
-                                            SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                        })
-                                })
-                                .catch(err => {
-                                    console.error(err)
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                })
-                            break
-                        case 'GetTokens':
-                            console.log('adding missing tokens to wallet assets.')
-                            SA.projects.foundations.utilities.httpRequests.getRequestBodyAsync(httpRequest, httpResponse)
-                                .then(body => {
-                                    let config = JSON.parse(body)
-                                    if (config.network === 'bsc') {
-                                        SA.projects.decentralizedExchanges.utilities.bsc.getTokens()
-                                            .then(response => {
-                                                SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(response), httpResponse)
-                                            })
-                                            .catch(err => {
-                                                console.error(err)
-                                                SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                            })
-                                    }
-                                })
-                                .catch(err => {
-                                    console.error(err)
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                })
-                    }
-                    break
-                case 'Social-Bots':
-                    switch (requestPath[2]) {
-                        case 'Discord-Test-Message':
-                            SA.projects.foundations.utilities.httpRequests.getRequestBodyAsync(httpRequest, httpResponse)
-                                .then(body => {
-                                    let config = JSON.parse(body)
-                                    let text = config.text
-                                    let socialBot = SA.projects.socialBots.botModules.discordBot.newSocialBotsBotModulesDiscordBot()
-                                    socialBot.initialize(config)
-                                        .then(response => {
-                                            console.log('httpInterface > Discord Bot >', response)
-                                            socialBot.sendMessage(text)
-                                                .then(response => {
-                                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(response), httpResponse)
-                                                })
-                                                .catch(err => {
-                                                    console.error(err)
-                                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                                })
-                                        })
-                                        .catch(err => {
-                                            console.error('error initializing discord bot', err)
-                                            SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                        })
-                                })
-                                .catch(err => {
-                                    console.error(err)
-                                })
-                            break
-                        case 'Slack-Test-Message':
-                            SA.projects.foundations.utilities.httpRequests.getRequestBodyAsync(httpRequest, httpResponse)
-                                .then(body => {
-                                    let config = JSON.parse(body)
-                                    let text = config.text
-                                    let socialBot = SA.projects.socialBots.botModules.slackBot.newSocialBotsBotModulesSlackBot()
-                                    socialBot.initialize(config)
-                                    socialBot.sendMessage(text)
-                                        .then(response => {
-                                            SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(response), httpResponse)
-                                        })
-                                        .catch(err => {
-                                            console.error(err)
-                                            SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                        })
-                                })
-                                .catch(err => {
-                                    console.error(err)
-                                })
-                            break
-                        case 'Twitter-Test-Message':
-                            SA.projects.foundations.utilities.httpRequests.getRequestBodyAsync(httpRequest, httpResponse)
-                                .then(body => {
-                                    config = JSON.parse(body)
-                                    let message = config.text
-                                    let socialBot = SA.projects.socialBots.botModules.twitterBot.newSocialBotsBotModulesTwitterBot(0)
-                                    socialBot.initialize(config)
-                                    socialBot.sendMessage(message)
-                                        .then(response => {
-                                            SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(response), httpResponse)
-                                        })
-                                        .catch(err => {
-                                            console.error(err)
-                                            SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                        })
-                                })
-                                .catch(err => {
-                                    console.error(err)
-                                })
-                            break
-                    }
-                    break
-                case 'Webhook': {
-                    switch (requestPath[2]) { // switch by command
-                        case 'Fetch-Messages': {
-                            let exchange = requestPath[3]
-                            let market = requestPath[4]
-
-                            /* Some validations */
-                            if (exchange === undefined) {
-                                console.log((new Date()).toISOString(), '[WARN] httpInterface -> Webhook -> Fetch-Messages -> Message with no Exchange received -> messageReceived = ' + messageReceived)
-                                SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                return
-                            }
-                            if (market === undefined) {
-                                console.log((new Date()).toISOString(), '[WARN] httpInterface -> Webhook -> Fetch-Messages -> Message with no market received -> messageReceived = ' + messageReceived)
-                                SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                return
-                            }
-
-                            let key = exchange + '-' + market
-
-                            let webhookMessages = webhook.get(key)
-                            if (webhookMessages === undefined) {
-                                webhookMessages = []
-                            }
-
-                            console.log((new Date()).toISOString(), '[INFO] httpInterface -> Webhook -> Fetch-Messages -> Exchange-Market = ' + exchange + '-' + market)
-                            console.log((new Date()).toISOString(), '[INFO] httpInterface -> Webhook -> Fetch-Messages -> Messages Fetched by Webhooks Sensor Bot = ' + webhookMessages.length)
-
-                            SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(webhookMessages), httpResponse)
-                            webhookMessages = []
-
-                            webhook.set(key, webhookMessages)
-                            break
-                        }
-                        case 'New-Message': {
-                            SA.projects.foundations.utilities.httpRequests.getRequestBody(httpRequest, httpResponse, processRequest)
-
-                            function processRequest(messageReceived) {
-                                if (messageReceived === undefined) {
-                                    return
-                                }
-
-                                let timestamp = (new Date()).valueOf()
-                                let source = requestPath[3]
-                                let exchange = requestPath[4]
-                                let market = requestPath[5]
-
-                                /* Some validations */
-                                if (source === undefined) {
-                                    console.log((new Date()).toISOString(), '[WARN] httpInterface -> Webhook -> New-Message -> Message with no Source received -> messageReceived = ' + messageReceived)
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                    return
-                                }
-                                if (exchange === undefined) {
-                                    console.log((new Date()).toISOString(), '[WARN] httpInterface -> Webhook -> New-Message -> Message with no Exchange received -> messageReceived = ' + messageReceived)
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                    return
-                                }
-                                if (market === undefined) {
-                                    console.log((new Date()).toISOString(), '[WARN] httpInterface -> Webhook -> New-Message -> Message with no market received -> messageReceived = ' + messageReceived)
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_FAIL_RESPONSE), httpResponse)
-                                    return
-                                }
-
-                                let key = exchange + '-' + market
-
-                                let webhookMessages = webhook.get(key)
-                                if (webhookMessages === undefined) {
-                                    webhookMessages = []
-                                }
-
-                                webhookMessages.push([timestamp, source, messageReceived])
-                                webhook.set(key, webhookMessages)
-
-                                console.log((new Date()).toISOString(), '[INFO] httpInterface -> Webhook -> New-Message -> Exchange-Market = ' + exchange + '-' + market)
-                                console.log((new Date()).toISOString(), '[INFO] httpInterface -> Webhook -> New-Message -> messageReceived = ' + messageReceived)
-                                console.log((new Date()).toISOString(), '[INFO] httpInterface -> Webhook -> New-Message -> Messages waiting to be Fetched by Webhooks Sensor Bot = ' + webhookMessages.length)
-                                SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_OK_RESPONSE), httpResponse)
-                            }
-
-                            break
-                        }
-                    }
-                }
-                    break
-                case 'Secrets': {
-                    switch (requestPath[2]) { // switch by command
-                        case 'Save-Singing-Accounts-Secrets-File': {
-                            SA.projects.foundations.utilities.httpRequests.getRequestBody(httpRequest, httpResponse, processRequest)
-
-                            async function processRequest(body) {
-                                try {
-
-                                    let filePath = global.env.PATH_TO_SECRETS + '/'
-                                    let fileName = "SigningAccountsSecrets.json"
-
-                                    SA.projects.foundations.utilities.filesAndDirectories.createNewDir(filePath)
-                                    SA.nodeModules.fs.writeFileSync(filePath + '/' + fileName, body)
-
-                                    console.log('[SUCCESS] ' + filePath + '/' + fileName + '  created.')
-
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(global.DEFAULT_OK_RESPONSE), httpResponse)
-
-                                } catch (err) {
-                                    console.log((new Date()).toISOString(), '[ERROR] httpInterface -> Secrets -> Save-Singing-Accounts-Secrets-File -> Method call produced an error.')
-                                    console.log((new Date()).toISOString(), '[ERROR] httpInterface -> Secrets -> Save-Singing-Accounts-Secrets-File -> err.stack = ' + err.stack)
-                                    console.log((new Date()).toISOString(), '[ERROR] httpInterface -> Secrets -> Save-Singing-Accounts-Secrets-File -> Params Received = ' + body)
-
-                                    let error = {
-                                        result: 'Fail Because',
-                                        message: err.message,
-                                        stack: err.stack
-                                    }
-                                    SA.projects.foundations.utilities.httpResponses.respondWithContent(JSON.stringify(error), httpResponse)
-                                }
-                            }
-                            break
-                        }
-                    }
-                }
-                    break
                 case 'Docs': {
                     switch (requestPath[2]) { // switch by command
                         case 'Save-Node-Schema': {
